tests/wpt/tests/IndexedDB/idbdatabase_createObjectStore.htm

<!DOCTYPE html>
<title>IDBDatabase.createObjectStore() - returns an instance of IDBObjectStore</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=resources/support.js></script>

<script>

var t = async_test(),
    open_rq = createdb(t)

open_rq.onupgradeneeded = function(e) {
    var db = e.target.result
    var objStore = db.createObjectStore('instancetest')

    assert_true(objStore instanceof IDBObjectStore, 'instanceof IDBObjectStore')
}

open_rq.onsuccess = function(e) {
    var db = e.target.result
    var objStore = db.transaction('instancetest', 'readonly', {durability: 'relaxed'}).objectStore('instancetest')

    assert_true(objStore instanceof IDBObjectStore, 'instanceof IDBObjectStore')
    t.done()
}
</script>

<div id=log></div>
